Rename segments.indent to segments.meta
