+++ src/sqlfluff/core/parser/segments/__init__.py
@@ -6,7 +6,7 @@ from sqlfluff.core.parser.segments.base import BaseSegment, UnparsableSegment
 from sqlfluff.core.parser.segments.generator import SegmentGenerator
 from sqlfluff.core.parser.segments.raw import RawSegment
 from sqlfluff.core.parser.segments.ephemeral import EphemeralSegment
-from sqlfluff.core.parser.segments.indent import Indent, Dedent, TemplateSegment
+from sqlfluff.core.parser.segments.meta import Indent, Dedent, TemplateSegment
 from sqlfluff.core.parser.segments.keyword import (
     KeywordSegment,
     SymbolSegment,
